\name{loops_help}
\alias{loops_help}

\title{
loops_help
}
\description{
This function gives a brief description of what is a while loop and/or a for loop
}
\usage{
loops_help(loop = loop)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{loop}{A string that specifies which loop you want a description for. Values can
  be either "while loop" or "for loop"
}
}

\author{
Rayane Prado Nunes
}
\note{
This function was developed for an assigment for the Pips course at the University of Amsterdam
}


