\name{vec_increaser}
\alias{vec_increaser}

\title{
vec_increaser
}
\description{
This function increases each element within a vector
}
\usage{
vec_increaser( variable = variable, by = 1, times = 1)
}

\arguments{
  \item{variable}{The numeric vector in which you want to increase each element
}
  \item{by}{The numeric value which specifies by how much you want to increase the elements each time.
  Can also be a vector as long as the lenght of the by vector matches the length of the variable vector
}
  \item{time}{The numeric value which specifies how many times you want the elements to be increased
}
}

\author{
Rayane Prado Nunes
}
\note{
This function was developed for an assigment for the Pips course at the University of Amsterdam
}
